requirements.txt
flask
tensorflow
numpy
matplotlib
Pillow
pdfkit
